requirements.txt
func_timeout
matplotlib
numpy
Pillow
pyserial
scipy
ttkbootstrap